migrations e model ok

+++ src/models/Account.model.ts
@@ -1,4 +1,4 @@
-import { Model, INTEGER, FLOAT } from 'sequelize';
+import { Model, INTEGER, DECIMAL } from 'sequelize';
 import db from '.';
 
 class Account extends Model {
@@ -18,7 +18,7 @@ Account.init(
     },
     balance: {
       allowNull: false,
-      type: FLOAT,
+      type: DECIMAL,
     },
   },
   {
